94 - Renames WFOProcessListPage to WFOProcessesListPage and moves logic to WFOProcessesList component
