+++ apps/wfo-ui/pages/processes/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { WFOProcessListPage } from '@orchestrator-ui/orchestrator-ui-components';
+import { WFOProcessesListPage } from '@orchestrator-ui/orchestrator-ui-components';
 
 export default function ProcessesPage() {
-    return <WFOProcessListPage />;
+    return <WFOProcessesListPage />;
 }
